+++ tests/codegen/service-codegen-tests/src/test/kotlin/com/test/utils.kt
@@ -80,9 +80,10 @@ internal fun cleanupService(proc: Process, gracefulWindow: Long = 5_000L) {
 
 private fun isWindows() = System.getProperty("os.name").lowercase().contains("windows")
 
-internal fun waitForPort(port: Int, timeoutSec: Long = 180): Boolean {
+internal fun waitForPort(port: Int, timeoutSec: Long = 180, proc: Process? = null): Boolean {
     val deadline = System.currentTimeMillis() + TimeUnit.SECONDS.toNanos(timeoutSec)
     while (System.currentTimeMillis() < deadline) {
+        proc?.inputStream?.bufferedReader()?.forEachLine { println(it) }
         try {
             Socket("localhost", port).use {
                 return true // Port is available
